[FEATURE] Added exceptions to check errors in Webtoon module.

+++ we-re-server/src/webtoons/webtoons.controller.ts
@@ -8,6 +8,7 @@ import {
   Delete,
   HttpCode,
   HttpStatus,
+  Logger,
 } from '@nestjs/common';
 import { WebtoonsService } from './webtoons.service';
 import { CreateWebtoonDto } from './dto/create-webtoon.dto';
@@ -27,6 +28,15 @@ import {
   ApiOperation,
   ApiTags,
 } from '@nestjs/swagger';
+import {
+  CustomBadTypeRequestException,
+  CustomNotFoundException,
+  CustomUnauthorziedException,
+} from 'src/utils/custom_exceptions';
+import {
+  stringToDays,
+  stringToProvidingCompany,
+} from 'src/entities/webtoon.entity';
 
 @ApiTags('Webtoons')
 @Controller('webtoons')
@@ -45,14 +55,19 @@ export class WebtoonsController {
   })
   @Get('detail/:id')
   async findOneDetailById(
-    @Param('id') id: string,
+    @Param('id') id: number,
   ): Promise<ReadWebtoonDetailDto> {
-    const result = await this.webtoonsService.findOneDetailById(+id);
-    result.storages = await this.storageService.findManyPublicListByWebtoonId(
-      +id,
-    );
-    result.reviews = await this.reviewService.findManyByWebtoonId(+id);
-    return result;
+    try {
+      if (!id) throw new CustomBadTypeRequestException('id', id);
+      const result = await this.webtoonsService.findOneDetailById(id);
+      result.storages = await this.storageService.findManyPublicListByWebtoonId(
+        id,
+      );
+      result.reviews = await this.reviewService.findManyByWebtoonId(id);
+      return result;
+    } catch (error) {
+      throw error;
+    }
   }
 
   @ApiOperation({
@@ -64,12 +79,17 @@ export class WebtoonsController {
   })
   @Get('list/liked/user/:userId')
   async findManyLikedThumbnailByUserId(
-    @Param('userId') userId: string,
+    @Param('userId') userId: number,
   ): Promise<ReadWebtoonThumbnailDto[]> {
-    // TODO:: 본인인지 체크 필요.
-    const { webtoonIds: ids } =
-      await this.likeService.findManyWebtoonIdsByUserId(+userId);
-    return this.webtoonsService.findManyThumbnailByIds(ids);
+    try {
+      if (!userId) throw new CustomBadTypeRequestException('userId', userId);
+      // TODO:: 본인인지 체크 필요.
+      const { webtoonIds: ids } =
+        await this.likeService.findManyWebtoonIdsByUserId(userId);
+      return this.webtoonsService.findManyThumbnailByIds(ids);
+    } catch (error) {
+      throw error;
+    }
   }
 
   @ApiOperation({
@@ -81,7 +101,11 @@ export class WebtoonsController {
   })
   @Get('list/new')
   findManyNewThumbnail(): Promise<ReadWebtoonThumbnailDto[]> {
-    return this.webtoonsService.findManyNewThumbnail();
+    try {
+      return this.webtoonsService.findManyNewThumbnail();
+    } catch (error) {
+      throw error;
+    }
   }
 
   @ApiOperation({
@@ -93,7 +117,11 @@ export class WebtoonsController {
   })
   @Get('list/hot')
   findManyHotThumbnail(): Promise<ReadWebtoonThumbnailDto[]> {
-    return this.webtoonsService.findManyHotThumbnail();
+    try {
+      return this.webtoonsService.findManyHotThumbnail();
+    } catch (error) {
+      throw error;
+    }
   }
 
   @ApiOperation({
@@ -105,17 +133,22 @@ export class WebtoonsController {
   })
   @Get('list/storage/:storageId')
   async findManyBreifInfoWithReviewByStorageId(
-    @Param('storageId') storageId: string,
+    @Param('storageId') storageId: number,
   ): Promise<ReadWebtoonBriefDto[]> {
-    // TODO:: NULL일때 Error 처리 필요.
-    const { webtoonIds: ids, userId } =
-      await this.storageService.findWebtoonIdListById(+storageId);
-    const result =
-      await this.webtoonsService.findManyBreifInfoWithReviewByStorageId(
-        ids,
-        userId,
-      );
-    return result;
+    try {
+      if (!storageId)
+        throw new CustomBadTypeRequestException('storageId', storageId);
+      const { webtoonIds: ids, userId } =
+        await this.storageService.findWebtoonIdListById(storageId);
+      const result =
+        await this.webtoonsService.findManyBreifInfoWithReviewByUserId(
+          ids,
+          userId,
+        );
+      return result;
+    } catch (error) {
+      throw error;
+    }
   }
 
   @ApiOperation({
@@ -131,10 +164,23 @@ export class WebtoonsController {
     @Param('day') day: string,
     @Param('providingCompany') providingCompany: string,
   ): Promise<ReadWebtoonThumbnailDto[]> {
-    return this.webtoonsService.findManyFilteredThumbnail(
-      day,
-      providingCompany,
-    );
+    try {
+      if (!stringToDays(day)) {
+        throw new CustomBadTypeRequestException('day', day);
+      }
+      if (!stringToProvidingCompany(providingCompany)) {
+        throw new CustomBadTypeRequestException(
+          'providingCompany',
+          providingCompany,
+        );
+      }
+      return this.webtoonsService.findManyFilteredThumbnail(
+        day,
+        providingCompany,
+      );
+    } catch (error) {
+      throw error;
+    }
   }
 
   @ApiOperation({ summary: 'create Webtoon' })
@@ -146,7 +192,11 @@ export class WebtoonsController {
   async createWebtoon(
     @Body() createWebtoonDto: CreateWebtoonDto,
   ): Promise<ReadWebtoonDetailDto> {
-    return await this.webtoonsService.createWebtoon(createWebtoonDto);
+    try {
+      return await this.webtoonsService.createWebtoon(createWebtoonDto);
+    } catch (error) {
+      throw error;
+    }
   }
 
   @ApiOperation({ summary: 'update Webtoon' })
@@ -159,15 +209,26 @@ export class WebtoonsController {
     @Param('id') id: number,
     @Body() updateWebtoonDto: UpdateWebtoonDto,
   ): Promise<ReadWebtoonDetailDto> {
-    return await this.webtoonsService.updateWebtoon(updateWebtoonDto);
+    try {
+      if (!id) throw new CustomBadTypeRequestException('id', id);
+      if (id !== updateWebtoonDto.id)
+        throw new CustomUnauthorziedException(`id is wrong.`);
+      return await this.webtoonsService.updateWebtoon(updateWebtoonDto);
+    } catch (error) {
+      throw error;
+    }
   }
 
   @ApiOperation({ summary: 'delete Webtoon' })
   @ApiNoContentResponse({ description: 'Request Success' })
   @HttpCode(HttpStatus.NO_CONTENT)
   @Delete(':id')
-  deleteReview(@Param('id') id: number): Promise<void> {
-    // 삭제 잘 되었다는 status code 반환~
-    return this.webtoonsService.deleteWebtoon(id);
+  async deleteWebtoon(@Param('id') id: number): Promise<void> {
+    try {
+      if (!id) throw new CustomBadTypeRequestException('id', id);
+      return await this.webtoonsService.deleteWebtoon(id);
+    } catch (error) {
+      throw error;
+    }
   }
 }
